잔디 깜

+++ 2022GameEngine1/Assets/Scripts/Road.cs
@@ -7,11 +7,11 @@ public class Road
 {
     public int _Num;            // 몇번째 길인지
     public Quaternion Rotation; // 돌아갔는지
-    public float _Roadlength;
+    public int _Roadlength;
     public GameObject Pref;
     public List<Obstacle> _Obstacles = new List<Obstacle>();
 
-    public Road(int num, float length, int minInterval, int maxInterval)     // 길 길이, 최소 간격, 최대 간격
+    public Road(int num, int length, int minInterval, int maxInterval)     // 길 길이, 최소 간격, 최대 간격
     {
         _Num = num;
         _Roadlength = length;
